ozw-rest: have heal and other controller operations
Note: we sort of screwed up the whole history. We forgot to stage for
commit the controller.py file, which actually implements the api that
shall be used here. However, pretty sure we've fixed bugs related to it
in the past few commits.

That said, from this point on the history should be sane.

Signed-off-by: Joao Eduardo Luis <joao@wipwd.dev>